+++ src/main/java/com/takku/project/controller/FundingManagementController.java
@@ -27,6 +27,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import com.takku.project.domain.FundingDTO;
 import com.takku.project.domain.ImageDTO;
 import com.takku.project.domain.ProductDTO;
+import com.takku.project.domain.StoreDTO;
 import com.takku.project.domain.UserDTO;
 import com.takku.project.service.FundingService;
 import com.takku.project.service.ImageService;
@@ -48,7 +49,7 @@ public class FundingManagementController {
 
 	@Autowired
 	ImageService imageService;
-	
+
 	// 내 펀딩 목록
 	@GetMapping
 	public String sellerFundings(@ModelAttribute("loginUser") UserDTO loginUser, Model model) {
@@ -64,77 +65,65 @@ public class FundingManagementController {
 		return "seller_fundings";
 	}
 
-	// 펀딩 입력 폼
-	@GetMapping("/new")
-	public String insertForm(@ModelAttribute("loginUser") UserDTO loginUser, Model model) {
-		Integer storeId = storeService.findStoreIdByUserId(loginUser.getUserId());
-
-		List<ProductDTO> productList = productService.selectProductByStoreId(storeId);
-		model.addAttribute("productList", productList);
-		return "seller_funding_add";
-	}
-
-	// 펀딩 등록 처리
+	// 펀딩 등록 처리(참고용임 지워야함)
 	@PostMapping
-	public String registerFunding( @ModelAttribute FundingDTO fundingDTO,
-	        @ModelAttribute("loginUser") UserDTO loginUser,
-	        @RequestParam("images") List<MultipartFile> images,
-	        HttpSession session,
-	        RedirectAttributes redirectAttributes) {
+	public String registerFunding(@ModelAttribute FundingDTO fundingDTO, @ModelAttribute("loginUser") UserDTO loginUser,
+			@RequestParam("images") List<MultipartFile> images, HttpSession session,
+			RedirectAttributes redirectAttributes) {
 
 		// 1. 로그인 유저로부터 storeId 조회
-	    Integer storeId = storeService.findStoreIdByUserId(loginUser.getUserId());
-	    fundingDTO.setStoreId(storeId);
+		Integer storeId = storeService.findStoreIdByUserId(loginUser.getUserId());
+		fundingDTO.setStoreId(storeId);
 
-	    // 2. 기본값 설정 (status: 준비중, currentQty: 0)
-	    fundingDTO.setStatus("준비중");
-	    fundingDTO.setCurrentQty(0);
+		// 2. 기본값 설정 (status: 준비중, currentQty: 0)
+		fundingDTO.setStatus("준비중");
+		fundingDTO.setCurrentQty(0);
 
-	    // 3. 펀딩 등록
-	    int result = fundingService.insertFunding(fundingDTO); // keyProperty로 fundingId 생성
+		// 3. 펀딩 등록
+		int result = fundingService.insertFunding(fundingDTO); // keyProperty로 fundingId 생성
 
-	    // 4. 파일 업로드 처리
-	    String uploadPath = session.getServletContext().getRealPath("/resources/images");
+		// 4. 파일 업로드 처리
+		String uploadPath = session.getServletContext().getRealPath("/resources/images");
 
-	    if (result > 0) {
-	        List<ImageDTO> imageDTOList = new ArrayList<>();
+		if (result > 0) {
+			List<ImageDTO> imageDTOList = new ArrayList<>();
 
-	        for (MultipartFile file : images) {
-	            if (!file.isEmpty()) {
-	                try {
-	                    // 고유 파일 이름 생성
-	                    String fileName = UUID.randomUUID() + "_" + file.getOriginalFilename();
-	                    String fullPath = uploadPath + File.separator + fileName;
+			for (MultipartFile file : images) {
+				if (!file.isEmpty()) {
+					try {
+						// 고유 파일 이름 생성
+						String fileName = UUID.randomUUID() + "_" + file.getOriginalFilename();
+						String fullPath = uploadPath + File.separator + fileName;
 
-	                    // 서버에 저장
-	                    file.transferTo(new File(fullPath));
+						// 서버에 저장
+						file.transferTo(new File(fullPath));
 
-	                    // 이미지 DTO 생성
-	                    ImageDTO imageDTO = new ImageDTO();
-	                    imageDTO.setFundingId(fundingDTO.getFundingId()); // FK 연결
-	                    imageDTO.setImageUrl("/resources/images/" + fileName);
+						// 이미지 DTO 생성
+						ImageDTO imageDTO = new ImageDTO();
+						imageDTO.setFundingId(fundingDTO.getFundingId()); // FK 연결
+						imageDTO.setImageUrl("/resources/images/" + fileName);
 
-	                    // DB 저장
-	                    imageService.insertImageUrl(imageDTO);
+						// DB 저장
+						imageService.insertImageUrl(imageDTO);
 
-	                    imageDTOList.add(imageDTO);
+						imageDTOList.add(imageDTO);
 
-	                } catch (IOException e) {
-	                    e.printStackTrace();
-	                    redirectAttributes.addFlashAttribute("resultMessage", "펀딩은 등록됐지만 이미지 업로드 중 오류 발생");
-	                    return "redirect:/seller/fundings";
-	                }
-	            }
-	        }
+					} catch (IOException e) {
+						e.printStackTrace();
+						redirectAttributes.addFlashAttribute("resultMessage", "펀딩은 등록됐지만 이미지 업로드 중 오류 발생");
+						return "redirect:/seller/fundings";
+					}
+				}
+			}
 
-	        // 5. DTO에 이미지 목록 추가 (필요시 뷰에서 활용)
-	        fundingDTO.setImages(imageDTOList);
-	        redirectAttributes.addFlashAttribute("resultMessage", "펀딩 등록 성공!");
-	    } else {
-	        redirectAttributes.addFlashAttribute("resultMessage", "펀딩 등록 실패");
-	    }
+			// 5. DTO에 이미지 목록 추가 (필요시 뷰에서 활용)
+			fundingDTO.setImages(imageDTOList);
+			redirectAttributes.addFlashAttribute("resultMessage", "펀딩 등록 성공!");
+		} else {
+			redirectAttributes.addFlashAttribute("resultMessage", "펀딩 등록 실패");
+		}
 
-	    return "redirect:/seller/fundings";
+		return "redirect:/seller/fundings";
 	}
 
 	// 펀딩 상세 조회
@@ -177,5 +166,133 @@ public class FundingManagementController {
 			redirectAttributes.addFlashAttribute("resultMessage", "삭제 실패");
 		}
 		return "redirect:/seller/fundings";
+	}
+
+	// 펀딩 만들기 -> 한정 상품 펀딩 or 일반 펀딩
+	@GetMapping("/create-step1")
+	public String selectStoreNameByUserId(Model model, HttpSession session) {
+		int userId = 3; // 임시 사용자 ID
+
+		StoreDTO store = storeService.selectStoreNameByUserId(userId);
+		model.addAttribute("store", store);
+
+		FundingDTO fundingDTO = new FundingDTO();
+
+		session.setAttribute("fundingDTO", fundingDTO);
+		session.setAttribute("store", store);
+		return "seller.createFunding";
+	}
+
+	// 상품 정보
+	@GetMapping("/create-step2")
+	public String createStep2(@RequestParam("type") String type, Model model, HttpSession session) {
+		// 상점이름
+		StoreDTO store = (StoreDTO) session.getAttribute("store");
+		model.addAttribute("store", store);
+
+		if ("general".equals(type)) {
+			return "pages/seller/create_normalFunding";
+		} else if ("limited".equals(type)) {
+			return "pages/seller/create_existMenu";
+		} else {
+			return "seller.createFunding";
+		}
+	}
+
+	// 펀딩이름 판매가 최소 판매개수 최대판대매수 인당구매
+	@PostMapping("/create-step3")
+	public String insertFundingMenuDetail(@ModelAttribute FundingDTO funding, HttpSession session, Model model) {
+
+		FundingDTO fundingDTO = (FundingDTO) session.getAttribute("fundingDTO");
+
+		StoreDTO store = (StoreDTO) session.getAttribute("store");
+		model.addAttribute("store", store);
+
+		fundingDTO.setFundingName(funding.getFundingName()); // 펀딩명
+		fundingDTO.setSalePrice(funding.getSalePrice()); // 판매가
+		fundingDTO.setTargetQty(funding.getTargetQty()); // 최소 판매 개수
+		fundingDTO.setMaxQty(funding.getMaxQty()); // 최대 판매 개수
+		fundingDTO.setPerQty(funding.getPerQty()); // 인당 구매 가능 개수
+
+		session.setAttribute("fundingDTO", fundingDTO);
+
+		return "pages/seller/create_insertDetail";
+	}
+
+	// ai, 직접입력 선택 창
+	@PostMapping("/create-step4")
+	public String writeType(@ModelAttribute FundingDTO funding, HttpSession session, Model model) {
+		StoreDTO store = (StoreDTO) session.getAttribute("store");
+
+		FundingDTO fundingDTO = (FundingDTO) session.getAttribute("fundingDTO");
+
+		fundingDTO.setStartDate(funding.getStartDate()); // 시작일
+		fundingDTO.setEndDate(funding.getEndDate()); // 마감일
+
+		session.setAttribute("fundingDTO", fundingDTO);
+		model.addAttribute("store", store);
+		return "pages/seller/select_writetype";
+	}
+
+	// 선택 후 제목, 내용 입력 창
+	@PostMapping("/create-step5")
+	public String handleWriteType(@RequestParam("type") String type, HttpSession session, Model model) {
+
+		StoreDTO store = (StoreDTO) session.getAttribute("store");
+		model.addAttribute("store", store);
+
+		if ("directly".equals(type)) {
+			return "pages/seller/funding_direct_input";
+		} else if ("ai".equals(type)) {
+			return "pages/seller/funding_ai_input";
+		}
+
+		// 잘못된 type 처리
+		return "redirect:/error";
+	}
+
+	@PostMapping("/submit-funding")
+	public String handleFundingBasicInfo(@ModelAttribute FundingDTO funding, HttpSession session) {
+		// 세션에서 기존 fundingDTO 가져오기
+		FundingDTO sessionDTO = (FundingDTO) session.getAttribute("fundingDTO");
+
+		// 만약 세션에 없으면 새로 생성 (예외 처리 목적)
+		if (sessionDTO == null) {
+			sessionDTO = new FundingDTO();
+		}
+
+		// 입력된 값만 덮어쓰기
+		sessionDTO.setFundingName(funding.getFundingName());
+		sessionDTO.setFundingDesc(funding.getFundingDesc());
+
+		/*
+		 * fundingService.insertFunding(sessionDTO);
+		 * 
+		 */
+
+		return "redirect:/seller/complete";
+	}
+
+	@GetMapping("/complete")
+	public String fundingComplete(HttpSession session, Model model) {
+		FundingDTO funding = (FundingDTO) session.getAttribute("fundingDTO");
+
+		session.removeAttribute("fundingDTO");
+
+		// 예외 처리 (없을 경우 홈으로)
+		if (funding == null) {
+			return "redirect:/seller/create-step1";
+		}
+
+		model.addAttribute("fundingName", funding.getFundingName());
+		model.addAttribute("startDate", funding.getStartDate());
+		// Date로 저장돼 있다면 포맷 필요
+		return "pages/seller/funding_complete";
+	}
+
+	// 기간 및 이미지
+	@GetMapping("/create-step3")
+	public String selectDateAndImage() {
+		return "pages/seller/create_insertDetail";
 	}
 }
